backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.7
langchain==0.0.350
elasticsearch==8.11.0
python-multipart==0.0.6
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3